requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Werkzeug==2.3.7
Pillow==10.0.1
python-socketio==5.8.0
celery==5.3.4
redis==5.0.1
psutil==5.9.6
python-dotenv==1.0.0